Use locale agnostic tolower() in dict.c hash function.

+++ redis-4.0-rc2/src/dict.c
@@ -1115,7 +1115,7 @@ void dictGetStats(char *buf, size_t bufsize, dict *d) {
 
 #include "sds.h"
 
-unsigned int hashCallback(const void *key) {
+uint64_t hashCallback(const void *key) {
     return dictGenHashFunction((unsigned char*)key, sdslen((char*)key));
 }
